+++ src/core/controllers/WebSocketController.ts
@@ -1,7 +1,7 @@
 import { DB, User } from "modelar";
 import { Controller } from "./Controller";
 import { WebSocket, Session } from "../tools/interfaces";
-import { config } from "../../init";
+import { config } from "../../index";
 
 export type WebSocketNextHandler = (controller: WebSocketController) => void;