+++ src/core/middleware/HttpDBHandler.ts
@@ -1,6 +1,6 @@
 import { App } from "webium";
 import { DB } from "modelar";
-import { config } from "../../init";
+import { config } from "../../index";
 import { Request, Response } from "../tools/interfaces";
 import { realDB } from "../tools/symbols";