+++ src/core/middleware/HttpStaticHandler.ts
@@ -1,7 +1,7 @@
 import serveStatic = require("serve-static");
 import * as fs from "fs";
 import { App } from "webium";
-import { config } from "../../init";
+import { config } from "../../index";
 
 export function handleStatic(app: App): void {
     if (config.statics) {