+++ src/core/middleware/WebSocketCookieHandler.ts
@@ -1,6 +1,6 @@
 import cookieParser = require("cookie-parser");
 import { Server } from "socket.io";
-import { config } from "../../init";
+import { config } from "../../index";
 import { WebSocket } from "../tools/interfaces";
 
 const parser = cookieParser(config.session.secret);