+++ src/index.ts
@@ -1,7 +1,26 @@
-import { config } from "./init";
+import * as date from "sfn-date";
+import * as Mail from "sfn-mail";
+import * as OutputBuffer from "sfn-output-buffer";
+import * as Logger from "sfn-logger";
+import * as Worker from "sfn-worker";
+import * as Validator from "sfn-validator";
+import Cache = require("sfn-cache");
 
-if (config.bluebird) {
-    global.Promise = require("bluebird");
-}
-
+export { date, Mail, OutputBuffer, Logger, Worker, Validator, Cache };
+export * from "sfn-scheduler";
+export * from "sfn-cookie";
+export * from "sfn-xss";
+export * from "./init";
+export * from "./core/tools/interfaces";
+export * from "./core/tools/functions";
+export * from "./core/tools/HttpError";
+export * from "./core/tools/SocketError";
+export * from "./core/tools/MarkdownParser";
+// load config before loading subsequent modules
+export * from "./core/bootstrap/ConfigLoader";
+export * from "./core/tools/Service";
+export * from "./core/tools/TemplateEngine";
+export * from "./core/tools/DevWatcher";
+export * from "./core/controllers/HttpController";
+export * from "./core/controllers/WebSocketController";
 export * from "./core/bootstrap/index";