+++ src/init.ts
@@ -1,48 +1,39 @@
-import "source-map-support/register";
 import * as path from "path";
 import * as fs from "fs";
-import { SFNConfig } from "./config";
-import { dirname } from 'path';
-export * from "./config";
+import "source-map-support/register";
+import { config as configEnv } from "dotenv";
 
 /** The version of framework. */
 export var version: string = require("../package.json").version;
 
-var _root = dirname(process.mainModule.filename);
+var appPath = path.dirname(process.mainModule.filename);
 
-/** The application path, usually it's the distribution path. */
-export const APP_PATH: string = (global["APP_PATH"] || _root).replace(/\\/g, "/");
+/** The root path of the project. */
+export const ROOT_PATH = path.normalize(appPath + "/..");
 
-/** The configuration of the program. */
-export var config: SFNConfig = Object.assign({}, SFNConfig);
+/**
+ * The application path, usually it's the distribution path. If your project 
+ * isn't written in TypeScript, then `APP_PATH` equals `SRC_PATH`.
+ */
+export const APP_PATH: string = appPath;
 
-if (fs.existsSync(APP_PATH + "/config.js")) {
-    // Load user-defined configurations.
-    let m = require(APP_PATH + "/config.js");
-    
-    if (typeof m.config === "object") {
-        config = Object.assign(config, m.config);
-    } else if (typeof m.env === "string") {
-        config = Object.assign(config, m);
-    }
-}
+var tsCfgFile = APP_PATH + "/../tsconfig.json";
+var tsCfg: { [x: string]: any; };
+var srcRoot: string = APP_PATH;
 
-var tscfgfile = APP_PATH + "/../tsconfig.json";
-var tscfg: { [x: string]: any; };
-var src_root: string = APP_PATH;
-
-if (fs.existsSync(tscfgfile)) {
+if (fs.existsSync(tsCfgFile)) {
     try {
-        tscfg = require(tscfgfile);
-        if (tscfg.compilerOptions && tscfg.compilerOptions.rootDir) {
-            src_root = path.normalize(APP_PATH + "/../" + tscfg.compilerOptions.rootDir);
-            src_root = src_root.replace(/\\/g, "/");
+        tsCfg = require(tsCfgFile);
+        if (tsCfg.compilerOptions && tsCfg.compilerOptions.rootDir) {
+            srcRoot = path.normalize(APP_PATH + "/../" + tsCfg.compilerOptions.rootDir);
         }
     } catch (e) { }
 }
 
-/** The source path. */
-export const SRC_PATH: string = src_root;
+/** The source code path. */
+export const SRC_PATH: string = srcRoot;
 
-/** Whether the program is running in development mode. */
-export var isDevMode = config.env == "dev" || config.env == "development";
+// support .env configuration file
+configEnv({
+    path: ROOT_PATH + "/.env"
+});